Creado service de voto, y DTO del objeto de la petición del voto

+++ backend/src/main/java/com/gym360/backend/repository/OfertaRepositorio.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfertaRepositorio extends JpaRepository<Oferta, Long> {
+public interface OfertaRepositorio extends JpaRepository<Oferta, Integer> {
